+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/ui/planets/PlanetsListFragment.kt
@@ -1,4 +1,4 @@
-package com.sw.sw_api_kotlin_project.ui.films
+package com.sw.sw_api_kotlin_project.ui.planets
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,59 +6,62 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.sw.sw_api_kotlin_project.adapters.FilmsAdapter
+import com.sw.sw_api_kotlin_project.adapters.PlanetsAdapter
 import com.sw.sw_api_kotlin_project.api.SWServiceClient
 import com.sw.sw_api_kotlin_project.api.liveData.SWApiLiveDataObserver
 import com.sw.sw_api_kotlin_project.base.BaseFragment
-import com.sw.sw_api_kotlin_project.data.model.Films
+import com.sw.sw_api_kotlin_project.data.model.Planet
 import com.sw.sw_api_kotlin_project.data.model.Results
-import com.sw.sw_api_kotlin_project.databinding.FragmentFilmsListBinding
-import com.sw.sw_api_kotlin_project.repository.FilmsRepository
+import com.sw.sw_api_kotlin_project.databinding.FragmentPeopleListBinding
+import com.sw.sw_api_kotlin_project.repository.PlanetsRepository
 import com.sw.sw_api_kotlin_project.utils.PageType
 
-class FilmsFragment : BaseFragment() {
-    private lateinit var viewModel: FilmsListViewModel
-    private var _binding: FragmentFilmsListBinding? = null
+
+class PlanetsListFragment : BaseFragment() {
+
+    private lateinit var viewModel: PlanetsViewModel
+    private var _binding: FragmentPeopleListBinding? = null
     private val binding get() = checkNotNull(_binding)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         viewModel = ViewModelProvider(
             this,
-            FilmsListViewModelFactory(FilmsRepository(SWServiceClient.getService()))
-        )[FilmsListViewModel::class.java]
+            PlanetsViewModelFactory(PlanetsRepository(SWServiceClient.getService()))
+        )[PlanetsViewModel::class.java]
     }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentFilmsListBinding.inflate(inflater, container, false)
+        _binding = FragmentPeopleListBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun initView() {
         super.initView()
         binding.nextButton.setOnClickListener {
-            getFilms(PageType.NEXT_PAGE)
+            getPlanets(PageType.NEXT_PAGE)
         }
         binding.previousButton.setOnClickListener {
-            getFilms(PageType.PREVIOUS_PAGE)
+            getPlanets(PageType.PREVIOUS_PAGE)
         }
-        getFilms(PageType.FIRST_PAGE)
+        getPlanets(PageType.FIRST_PAGE)
     }
 
-    private fun getFilms(pageType: PageType) {
-        val filmsObserver = object : SWApiLiveDataObserver<Results<Films>>() {
-            override fun onSuccess(data: Results<Films>?) {
-                val films = data!!
+    private fun getPlanets(pageType: PageType) {
+        val planetsObserver = object : SWApiLiveDataObserver<Results<Planet>>() {
+            override fun onSuccess(data: Results<Planet>?) {
+                val planets = data!!
                 binding.progressBar.visibility = View.GONE
                 binding.recyclerView.visibility = View.VISIBLE
                 binding.previousButton.visibility = View.VISIBLE
                 binding.nextButton.visibility = View.VISIBLE
-                binding.previousButton.isEnabled = films.previous != null
-                binding.nextButton.isEnabled = films.next != null
-                val adapter = FilmsAdapter(films.results)
+                binding.retryButton.visibility = View.GONE
+                binding.previousButton.isEnabled = planets.previous != null
+                binding.nextButton.isEnabled = planets.next != null
+                val adapter = PlanetsAdapter(planets.results)
                 binding.recyclerView.adapter = adapter
                 binding.recyclerView.layoutManager = LinearLayoutManager(context)
             }
@@ -68,8 +71,8 @@ class FilmsFragment : BaseFragment() {
                 binding.errorText.visibility = View.VISIBLE
                 binding.previousButton.visibility = View.GONE
                 binding.nextButton.visibility = View.GONE
-                binding.errorText.text = errorMessage
                 binding.retryButton.visibility = View.VISIBLE
+                binding.errorText.text = errorMessage
             }
 
             override fun onLoading() {
@@ -81,7 +84,8 @@ class FilmsFragment : BaseFragment() {
                 binding.retryButton.visibility = View.GONE
             }
         }
-        viewModel.getFilms(pageType).observe(viewLifecycleOwner, filmsObserver)
+
+        viewModel.getPlanets(pageType).observe(viewLifecycleOwner, planetsObserver)
     }
 
     override fun onDestroy() {
